move tools to where they belong
